+++ app/Command/XCat.php
@@ -272,22 +272,24 @@ class XCat
 
     public function setTelegram()
     {
-        // $bot = new BotApi(Config::get('telegram_token'));
-        // $ch = curl_init();
-        // curl_setopt($ch, CURLOPT_URL, sprintf('https://api.telegram.org/bot%s/deleteWebhook', Config::get('telegram_token')));
-        // curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-        // curl_setopt($ch, CURLOPT_CONNECTTIMEOUT, 5);
-        // $deleteWebhookReturn = json_decode(curl_exec($ch));
-        // curl_close($ch);
-        // if ($deleteWebhookReturn->ok && $deleteWebhookReturn->result && $bot->setWebhook(Config::get('baseUrl') . '/telegram_callback?token=' . Config::get('telegram_request_token')) == 1) {
-        //     echo ('设置成功！' . PHP_EOL);
-        // }
-
-        $WebhookUrl = (Config::get('baseUrl') . '/TelegramCallback?token=' . Config::get('telegram_request_token'));
-        $telegram = new \Telegram\Bot\Api(Config::get('telegram_token'));
-        $telegram->removeWebhook();
-        if ($telegram->setWebhook(['url' => $WebhookUrl])) {
-            echo('Bot @' . $telegram->getMe()->getUsername() . ' 设置成功！');
+        if (Config::get('use_new_telegram_bot') === true) {
+            $WebhookUrl = (Config::get('baseUrl') . '/TelegramCallback?token=' . Config::get('telegram_request_token'));
+            $telegram = new \Telegram\Bot\Api(Config::get('telegram_token'));
+            $telegram->removeWebhook();
+            if ($telegram->setWebhook(['url' => $WebhookUrl])) {
+                echo('New Bot @' . $telegram->getMe()->getUsername() . ' 设置成功！');
+            }
+        } else {
+            $bot = new BotApi(Config::get('telegram_token'));
+            $ch = curl_init();
+            curl_setopt($ch, CURLOPT_URL, sprintf('https://api.telegram.org/bot%s/deleteWebhook', Config::get('telegram_token')));
+            curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
+            curl_setopt($ch, CURLOPT_CONNECTTIMEOUT, 5);
+            $deleteWebhookReturn = json_decode(curl_exec($ch));
+            curl_close($ch);
+            if ($deleteWebhookReturn->ok && $deleteWebhookReturn->result && $bot->setWebhook(Config::get('baseUrl') . '/telegram_callback?token=' . Config::get('telegram_request_token')) == 1) {
+                echo ('Old Bot 设置成功！' . PHP_EOL);
+            }
         }
     }
