+++ app/Utils/Telegram/TelegramTools.php
@@ -2,12 +2,48 @@
 
 namespace App\Utils\Telegram;
 
-use App\Models\TelegramTasks;
+use App\Models\{TelegramTasks, User};
 use App\Services\Config;
 use App\Utils\Tools;
 
 class TelegramTools
 {
+
+    /**
+     * 搜索用户
+     *
+     * @param string $value  搜索值
+     * @param string $method 查找列
+     * 
+     * @return \App\Models\User
+     */
+    public static function getUser($value, $method = 'telegram_id')
+    {
+        return User::where($method, $value)->first();
+    }
+
+    /**
+     * Sends a POST request to Telegram Bot API.
+     * 伪异步，无结果返回.
+     *
+     * @param array $params
+     *
+     * @return string
+     */
+    public static function SendPost($Method, $Params)
+    {
+        $URL = 'https://api.telegram.org/bot' . Config::get('telegram_token') . '/' . $Method;
+        $POSTData = json_encode($Params);
+        $C = curl_init();
+        curl_setopt($C, CURLOPT_URL, $URL);
+        curl_setopt($C, CURLOPT_POST, 1);
+        curl_setopt($C, CURLOPT_HTTPHEADER, ['Content-Type:application/json; charset=utf-8']);
+        curl_setopt($C, CURLOPT_POSTFIELDS, $POSTData);
+        curl_setopt($C, CURLOPT_TIMEOUT, 1);
+        curl_exec($C);
+        curl_close($C);
+    }
+
     /**
      * 用户识别搜索字段
      *
