+++ config/routes.php
@@ -59,12 +59,13 @@ $app->get('/500', App\Controllers\HomeController::class . ':page500');
 $app->post('/notify', App\Controllers\HomeController::class . ':notify');
 $app->get('/tos', App\Controllers\HomeController::class . ':tos');
 $app->get('/staff', App\Controllers\HomeController::class . ':staff');
-$app->post('/telegram_callback', App\Controllers\HomeController::class . ':telegram');
-$app->post('/tomato_back/{type}', 'App\Services\Payment:notify');
-$app->get('/tomato_back/{type}', 'App\Services\Payment:notify');
+$app->post('/tomato_back/{type}', App\Services\Payment::class . 'notify');
+$app->get('/tomato_back/{type}', App\Services\Payment::class . 'notify');
 
 // New Telegram
 $app->post('/TelegramCallback', App\Controllers\HomeController::class . ':NewTelegram');
+// Old Telegram
+$app->post('/telegram_callback', App\Controllers\HomeController::class . ':telegram');
 
 // User Center
 $app->group('/user', function () {
